Change name to udi_interface.
So that this new API may co-exist with the legacy API defined for
Polyglot version 2. To reduce the effort to port some node servers
to Polyglot version 3, the existing interface API has been made
available as polyinterface version 3.0.x. The changes between
polyinterface 2.x and 3.0.x are minimal.  However, it does not
take advantage of new feature in Polyglot version 3 nor does if fix
existing issues with the interface API.

This new interface API, udi_interface, is now the offical API that
should be used for new node server development going forward.