Add new event callbacks and modify node start method behavior
Added new event callbacks:
  - onStartDone() that calls the callback when a node's start method
    has finished.
  - onAddNodeDone() that calls the callback when polyglot has finished
    adding a node.

The onStart() callback has been changed to take both a node address
and callback function. The callback will only be called when the
node with the matching address has been added.

Changed the behavior for calling the node's start method.  Instead
of calling it at the end of the initial config response.  We now
call it when polyglot has notified us that the node has been added.
We also wrap the callback so that we can keep track of when the start
method has finished and trigger the onStartDone callback added above.

Signed-off-by: Bob Paauwe <bpaauwe@yahoo.com>