Fix setDriver & reportDriver
It looks like setDriver and reportDriver were taken from the polyglot
cloud interface implementation where the driver array from the node
server is converted to a dictionary to use internally.  But the
conversion to a dictionary was never done.

Given that the driver array is something created by the node server,
it seems best to not mess with the structure internally as that can
confuse node server authors.  Re-implement both methods based on
the PG2 implementation.