Build initial node list from the nodes returned in the config.
The config includes node information for all the nodes stored in
the database.  Use that information to build the internal list of
nodes.

If possible, create the node using the node server's custom class.
However, if the class adds additional parameters to the initialization
it won't be able to.  The node server should call addNode on these
nodes to get the proper class objects in the internal node list.