Queue most message types.
When messages are recieived over the MQTT connection, we should
handle them as quickly as possible. For some messages types,
like getAll, the time it takes to process is long enough that
messages can get dropped.

So simply queue most messages and let the input handling thread
process them. The exception is messages that we don't really handle
but might just log that they were recieved.