Fix processing of nodes in config.
When PG3 sends the configuration, that includes the list of nodes
as stored in the database.  This list is not node class objects.

Previously, we would try to create a node class object from this
information, but that results in creating a duplicate which can
lead to bad behavior, exspecally if two controller nodes are running.

Instead, we should just save off the information in the _nodes list
as a list of dictionaries. We now only use this information internally
to update the node object once it has been created.

We also maintain a nodes list which is a list of the actual node
class objects.  This is what is used to access node methods like
runcmd().  Nodes are only added to this list when the node server
calls addNode().

It's not possible to automatically recreate the node list in the
interface from the configuration data so don't try.