Add some handshaking with interface.
If the interface starts asking PG3 for information before the node
server has registered callbacks to handle processing the information,
there will be problems.

Previously, the interface would send the config and getAll messages
to PG3 right after subscribing to the communication channel topic.
This would lead to random behavior as sometimes the reponse message
would come before the callbacks had been registered.

The solution is to add a 'ready' method in the interface that the
node server can call after it has registered the required callbacks.
This then triggers the interface to send the config and getAll messages
to PG3.