logger: separate interface logging control from NS logging
When the interface debug logs are enabled, the volume of log
messages generated is very high and can make it difficult to
track down isues within the node server itself.

One the interface code is stabel, it should only be necessary
to enable debug level logs in the interface when trying to track
down an issue related to node server/interface interactions.

Change the behavior so that log level settings, by the user, using
the dashboard's log level selector, change the node server's log
level by default.

Add a notifier that the node server can subscribe to that is
triggered whenever the log level is changed.  This allows the node
server to override what the user has selected and trigger additional
logging in the interface modules if so desired.