Add new event called config_done.
This event is published after PG3 has sent all the configuration
info and the interface has published that information to the node
server.

The node server would use this event to know when it is safe to
start processing the configuration data recieved when it depends
on multiple types of configuration data (i.e. customparams and config)

Because data is only sent if it exist in the PG3 database, there
are cases when the node server will not recieve an event for
data.  I.E. if there are no custom parameters yet, the node server
will not recieive a CUSTOMPARAMS event.  If you get a CONFIGDONE
event but have not received a CUSTOMPARAMS event, you know you will
not be getting a CUSTOMPARAMS event unless the user updates them.