nsdata custom data
Added a new custom data key 'nsdata'.  This data is pulled from
the node server store database and passed on to the node server
via the CUSTOMNS event.

PG3 will check for changes to this data and send the updated data
to the node server in a event.

Example usage might be a secret key that may need to be updated if
the original is revoked. By setting the secret key in nsdata, it will
be sent to the node server when first installed and the developer can
update the key as needed.  Every update will be pulled by all running
installations of the node server.