Delay CONFIGDONE event until config is really finished.
Config consist of the config data plus all of the custom data.  Since
these are all separate events and it may take time for the node server
to process the events.  We need to wait until all of the initial config
related events have been sent AND processed before sending the
CONFIGDONE event.