Properly handle custom database update responses.
PG3 will now use the 'custom' key to indicate the response is
from updating the custom database.

Previously, it would respond using the custom key.  Since we allow
node servers to define custom key, we had no way to know that
what the custom key represented and would log this as an invalid
command from PG3.  We can now see that this is a custom database
update and respond correctly in the interface.