setController / addNode
Add an interface API method, setController.  This takes both a
node address and a driver name.  This is saved to the PG3 database
and PG3 will set the driver specified of the node specified to the
connection status (0, 1, 2) representing disconnected, connected, and
failed.

The addNode() method is updated to take a optional argument, conn_status
that, if set, will call setController to set the newly added noded as the
node (with the specified driver) to use for connection status.