Update driver defaults during node initialization.
Since we save the changes to node drivers in the PG3 database, lets
make use of that and restore those changes to the node drivers when
the nodes are initialized.

This should eliminate the status changes that could occur when a
node server was started.

Prevously, when the node server started, it would send the default
driver values to the ISY. Typically these were all zero.  This would
change the ISY driver statuses.  The node server would then need to get
the current device values and update the ISY.

Now, when the node server starts, it will send the last driver values
which should match what the ISY already has, and thus, not cause a
status change.

There is one caveat, if a controller node is created, it is typically
done before the database info is sent so the controller node will
continue to behave as before.