Clean up internal and external node lists representations.
Internally, the self._nodes list is now just the list of nodes
as saved in the database. This currently comes from the config
structure/message from PG3.

self.internal_nodes is the list of nodes added by the node server.
This is a list of the actual node class objects.

db_getDrivers has been deprecated as part of the API (still used
internally) with a new function: getDriversFromDb that is now
exposed to the node server so that it follow the nameing convention
used for Interface methods.

getDriversFromDb supports an addition use case where you can pass
an array of node address and get back the matching list of node data
as stored in the database.

Improved the documentation for this method.