SSL certificate and connection status.
Add support to use a client certificate for a secure mqtt
connection.  This assumes the certificates are in the node server's
directory and are named with the node server's ID (uuid_slot).

Add support for publishing the client connection status when we
connect.  This is to notify PG3 that we have connected to the mqtt
broker.

Publish a client disconnect status before we disconnect from the
mqtt broker.