Change address filter behavior for non-start events.
Event subscriptions take an optional parameter specifying and address
filter.  The purpose of this filter is to only call the event handler
if the address matches the address specified when the event is published.

However, only the START event includes an address when published, all
other events set the address to None. The result is that if you specify
an address when subscribing to any event except START, your subscription
will never trigger and never call the event handler because non-None
will never equal None.

This wasn't really the intended behavior. Events published with address
set to None were ment to apply to any/all address. This change ignores
the filter address in the subscription if the the event is published with
address=None.

Basically all event subscriptions will trigger the handler except for
START now.