Fix node rename operations
When we call PG3 to rename a node it renames the node in it's
database and on the ISY.  However, it also needs to update the
node class object in the interface.

There are also some restrictions.  We shouldn't rename the node
in the database or ISY if the node class object hasn't been created
yet.

While we're at it make sure that addNode warns about renaming the
node properly.

Issue: #121