Add rename option to addNode
addNode now takes an optional parameter; rename that can be
either true or false.  It defaults to false.

If set to true, it will cause the node to be renamed on the ISY
and in the PG3 database if using PG3 version 3.1.15 or later.

Earlier versions of PG3 should ignore this parameter