Add support for node server optional feature support
The first optional feature is support for requestId. If the
node server advertises that it supports requestId then it will
send the report response after query, status, and command. If
it does not advertise this support, PG3 will continue to
autorespond.

This builds on top of the existing code that allows the node server
to send it's version number.