Use a queue to handle messages that are sent to PG3
Instead of publishing messages directly to PG3, queue the messages
in a queue and then run a thread that pulls from the queue and
publishes them one-by-one.

Because a node server is multi-threads, there was a possiblity
that two threads would try to publish messages at about the same
time and one of them would be corrupted, causing SSL errors.