re-write pub class (publish/subscribe)
The pub class has a couple of issues that can lead to undesired
resource use.

We were trying to solve the problem that ocured when nodes
would subscribe to events, that typically only happen once
during initialization, after those events were origanlly
published.

To do that, we saved each event as it was published and then
re-published them when a new node subscribed.

This was fine for events that are published once or at most a
a dozen times.  However for events like POLL and NSINFO which
can happen multiple times a minute the number of events saved
adds up quickly.  This causes the memory use for saved events
to constantly grow. Also, if a node is created days or weeks
after the node server starts and then subscribes to POLL
events, it will get inudated with thousands (millions?) of
events and possibly exhaust the thread pool.

A better solution is to only save the last event of each type.
This way, we save at most 22 events and memory use for saved
events will not grow over time.  It also means that a new node
will only get the latest event and not thousands+ events if
it subscribes after the node server has been running a while.