Use a single thread to publish messages to PG3.
The MQTT client is not thread safe.  Publishing messages from
multiple threads can cause issues down in the SSL layer which
leads to a exception and eventually, the MQTT connection being
disconnected.

To prevent this, create a message queue and a thread that
monitors the queue, sending the messages in the queue to PG3
one at a time.

This will obviously have a performance impact but under testing
with a verbose node server, the queue never grew beyond 12 which
is emptied in a very short period of time.

The queue thread is started in the mqtt on_connect handler and
stopped in the on_disconnect handler. This should prevent it
from running more than one thread.