Switch to using wait_for_publish()
Instead of delaying for a fixed period of time, use
wait_for_publish() to make sure the message is published before
sending the next message.

The goal is to send out messages as fast as possible but also
make sure we don't have messages overlapping as we believe that
if the messages overlap, it can result in an SSL handshake error.